Simplify server to single route.
