+++ examples/duplex-speech-streaming-vite-react-fastify/package.json
@@ -4,8 +4,8 @@
   "license": "MIT",
   "private": true,
   "scripts": {
-    "dev-client": "vite",
-    "dev-server": "tsx src/server/server.ts",
+    "client": "vite",
+    "server": "tsx src/server/server.ts",
     "build": "tsc && vite build",
     "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "preview": "vite preview"
