+++ examples/duplex-speech-streaming-vite-react-fastify/src/eventSchema.ts
@@ -9,7 +9,4 @@ export const eventSchema = z.discriminatedUnion("type", [
     type: z.literal("speech-chunk"),
     base64Audio: z.string(),
   }),
-  z.object({
-    type: z.literal("finish"),
-  }),
 ]);
