--- examples/duplex-speech-streaming-vite-react-fastify/src/server/EndpointRun.ts
@@ -1,19 +1,0 @@
-import { AsyncQueue, DefaultRun } from "modelfusion";
-
-export class EndpointRun<EVENT> extends DefaultRun {
-  readonly eventQueue: AsyncQueue<EVENT> = new AsyncQueue();
-  readonly endpointName: string;
-
-  constructor({ endpointName }: { endpointName: string }) {
-    super();
-    this.endpointName = endpointName;
-  }
-
-  publishEvent(event: EVENT) {
-    this.eventQueue.push(event);
-  }
-
-  finish() {
-    this.eventQueue.close();
-  }
-}
